#5210 refactoring: inout line validation in a new repo
https://github.com/metasfresh/metasfresh/issues/5210
